Change datatype

+++ features/merchant/shop_page/src/main/java/com/tokopedia/shop/info/view/viewmodel/ShopInfoViewModel.kt
@@ -98,7 +98,7 @@ class ShopInfoViewModel @Inject constructor(
         }
     }
 
-    fun getShopGoApotikData(shopId: Int, warehouseId: Int) {
+    fun getShopGoApotikData(shopId: Long, warehouseId: Long) {
         launchCatchError(block = {
             coroutineScope {
                 val shopEpharmData = withContext(coroutineDispatcherProvider.io) {
